Add QemuCmdDomainConfigurator to converter
Extract logic responsible for configuring guest's QEMU command and arguments to
a dedicated QemuCmdDomainConfigurator in the compute package.
