+++ pkg/virt-launcher/virtwrap/converter/compute/BUILD.bazel
@@ -16,6 +16,7 @@ go_library(
         "launch_security.go",
         "os.go",
         "panic_devices.go",
+        "qemu_cmd.go",
         "rng.go",
         "sound.go",
         "sysinfo.go",
@@ -30,6 +31,7 @@ go_library(
         "//pkg/config:go_default_library",
         "//pkg/container-disk:go_default_library",
         "//pkg/downwardmetrics:go_default_library",
+        "//pkg/ignition:go_default_library",
         "//pkg/pointer:go_default_library",
         "//pkg/tpm:go_default_library",
         "//pkg/util:go_default_library",
