+++ pkg/virt-launcher/virtwrap/util/BUILD.bazel
@@ -16,7 +16,7 @@ go_library(
         "//pkg/virt-handler/cgroup:go_default_library",
         "//pkg/virt-launcher/virtwrap/api:go_default_library",
         "//pkg/virt-launcher/virtwrap/cli:go_default_library",
-        "//pkg/virt-launcher/virtwrap/converter:go_default_library",
+        "//pkg/virt-launcher/virtwrap/converter/compute:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/log:go_default_library",
         "//vendor/golang.org/x/sys/unix:go_default_library",
